refactor: move ecstore rebalance state impls
